Try building this with the ability to sort stuff from ats2-floattypes.
(There are BTW rumors on the net that ‘long double’ is quad
precision. I believe this is not so with GCC on x86-64, despite that
such a double is stored in 16 bytes. The 16-byte storage is specified
by the x86-64 ABI. To get quad precision from this you need an option,
but it is better to use _Float128.)

--HG--
branch : with-ats2-floattypes